App/requirements.txt
alabaster==0.7.12
opencv-python-headless== 4.2.0.34
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
gunicorn==20.0.4
json5==0.8.5
jsonschema==3.0.2
numpy==1.16.5
numpydoc==0.9.1
Pillow==6.2.0
pipenv==2018.11.26
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
simple-http-server==0.1.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
sqlparse==0.3.0
srsly==1.0.2
whitenoise==5.0.1